Migration from Material design to Bootstrap 4
The material design assets used was not on Toolforge and
the UI of the app was messed up. This caused a migration
from MD to Bootstrap 4 which is supported by Toolforge.

This patch also solves the ticket here: T172609, so that
the app no longer fetches content out of Toolforge.

Bug: T172609.
Signed-off-by: Alangi Derick <alangiderick@gmail.com>